prevent mocked child_process from leaking into other tests
